Godrej Panorama Kolshet Road Thane
Godrej Panorama is an upcoming residential development by Godrej Properties in the prime location of Kolshet Road thane. This development is in the city which is famous as the IT center of the nation and is soon organizing itself as a famous real estate property across the country. This latest construction by the famous developer Godrej Group has been considered and established by using the latest and modern technologies.

The project provides 1, 2 & 3 BHK Premium Residences with great facilities and communities. Godrej Panorama Kolshet Road Thane Built elegantly and charmingly, the awe-inspiring homes are developed in luxurious green surroundings that are inspirational and attractive. The game facility, 80% green open area, spa, Reflexology Track, swimming pool, gymnasium, 2 Luxurious Grand Clubs, and a committed child's playground are some of the significant highlights of the new residential venture.

For More Details :
Visit - https://godrej.developerprojects.in/upcoming-projects-in-thane/godrej-panorama-kolshet-road/